cybersec-copilot/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
requests
scikit-learn
pandas
numpy
joblib